#38 Particle Effects: - Refactored effects class to contain sub objects and existing effects class to be a postprocessing subclass - Adding ship thruster from older codebase - Moved debugging section of main menu to bottom
