+++ client/src/app/scenes/base.js
@@ -29,7 +29,10 @@ export default class SceneBase {
        * 
        * @memberof Object { postprocessing.EffectComposer }
        */
-      this.effects = false;
+      this.effects = {
+        particles: false,
+        postprocessing: false
+      };
 
       /**
        * Exit sign.
