+++ client/src/app/ui/main_menu.js
@@ -14,6 +14,79 @@ export default class Main_Menu {
             expanded: true
         });
 
+        const exit_game = this.pane.addButton({
+            title: 'Exit to Main Menu',
+            hidden: true
+        });
+        exit_game.on('click', () => {
+            console.log('Exit to Main Menu, closing game session');
+            window.l.controls.deactivate()
+            window.l.current_scene.ui.hide_flight_instruments();
+
+            // Show game mode buttons.
+            single_player.hidden = false;
+            multi_player.hidden = false;
+
+            // Hide game exit button to return to main menu.
+            exit_game.hidden = true;
+
+            if (window.l.multiplayer.connected) {
+                window.l.multiplayer.disconnect();
+            }
+
+        });
+
+        const single_player = this.pane.addButton({
+            title: 'Single Player',
+        });
+        single_player.on('click', () => {
+            console.log('Single player launched');
+            window.l.controls.activate();
+            window.l.current_scene.ui.show_flight_instruments();
+
+            // Hide game mode buttons.
+            single_player.hidden = true;
+            multi_player.hidden = true;
+
+            // Show game exit button to return to main menu.
+            exit_game.hidden = false;
+        });
+
+        const multi_player = this.pane.addButton({
+            title: 'Multi Player',
+        });
+        multi_player.on('click', () => {
+            console.log('Multi player launched');
+            window.l.controls.activate();
+            window.l.current_scene.ui.show_flight_instruments();
+
+            // Hide game mode buttons.
+            single_player.hidden = true;
+            multi_player.hidden = true;
+
+            // Show game exit button to return to main menu.
+            exit_game.hidden = false;
+
+            let serverLocation = window.l.env =='Dev' ? 'lcl.langenium.com:8090' : 'test.langenium.com:42069' ;
+
+            window.l.multiplayer.connect('//' + serverLocation);
+
+        });
+
+        const settings = this.pane.addButton({
+            title: 'Settings',
+        });
+        settings.on('click', () => {
+            console.log('Settings launched');
+        });
+
+        const help = this.pane.addButton({
+            title: 'Help',
+        });
+        help.on('click', () => {
+            console.log('Help launched');
+        });
+
         if ( window.l.current_scene.debug ) {
             const debugging = this.pane.addFolder({
                 title: 'Debugging',
@@ -89,79 +162,6 @@ export default class Main_Menu {
                 interval: 200
             })
         }
-
-        const exit_game = this.pane.addButton({
-            title: 'Exit to Main Menu',
-            hidden: true
-        });
-        exit_game.on('click', () => {
-            console.log('Exit to Main Menu, closing game session');
-            window.l.controls.deactivate()
-            window.l.current_scene.ui.hide_flight_instruments();
-
-            // Show game mode buttons.
-            single_player.hidden = false;
-            multi_player.hidden = false;
-
-            // Hide game exit button to return to main menu.
-            exit_game.hidden = true;
-
-            if (window.l.multiplayer.connected) {
-                window.l.multiplayer.disconnect();
-            }
-
-        });
-
-        const single_player = this.pane.addButton({
-            title: 'Single Player',
-        });
-        single_player.on('click', () => {
-            console.log('Single player launched');
-            window.l.controls.activate();
-            window.l.current_scene.ui.show_flight_instruments();
-
-            // Hide game mode buttons.
-            single_player.hidden = true;
-            multi_player.hidden = true;
-
-            // Show game exit button to return to main menu.
-            exit_game.hidden = false;
-        });
-
-        const multi_player = this.pane.addButton({
-            title: 'Multi Player',
-        });
-        multi_player.on('click', () => {
-            console.log('Multi player launched');
-            window.l.controls.activate();
-            window.l.current_scene.ui.show_flight_instruments();
-
-            // Hide game mode buttons.
-            single_player.hidden = true;
-            multi_player.hidden = true;
-
-            // Show game exit button to return to main menu.
-            exit_game.hidden = false;
-
-            let serverLocation = window.l.env =='Dev' ? 'lcl.langenium.com:8090' : 'test.langenium.com:42069' ;
-
-            window.l.multiplayer.connect('//' + serverLocation);
-
-        });
-
-        const settings = this.pane.addButton({
-            title: 'Settings',
-        });
-        settings.on('click', () => {
-            console.log('Settings launched');
-        });
-
-        const help = this.pane.addButton({
-            title: 'Help',
-        });
-        help.on('click', () => {
-            console.log('Help launched');
-        });
 
         return this.pane;
     }
